Create Resolution of Problem 127 of Project Euler
For this problem I first created an array of primes until 120k, using a sieve of Eratosthenes,
then I computed the factorization of all the numbers until 120k, then I made the pr-computation
of all the rad values until 120k, to avoid over computation of the same numbers.

After that, because I noticed that there is only possible to have rad(a.b.c)<c if there is only one
prime number among a,b,c, I checked if among a,b,c during the for cycle, c is a prime number,
if yes, I would continue the tests, for the next pair (a,b).

If it passed this test, I would make the calculation of rad(a.b.c).
As gcd(a,b)=1 then it is guaranteed that rad(a.b.c)= because they are co-prime numbers, so I
would only made the calculation of rad(a).rad(b).rad(c) instead of the heavier rad(a.b.c). If this
would pass the test, then I would at last, test if gcd(a,b)=1. Because c=a+b then
gcd(a,b)=gcd(a,a+b)=gcd(b,b+c)=1 is the equivalent to gcd(a,b)=1, lighter in the calculation.

So for this problem:

    pr-computation of all the prime numbers until 120k.
    pr-computation of all the factorizations until 120k.
    pr-computation of all the rad until 120k.
    test if c is a prime number, if yes, continue to the next pair (a,b).
    test if rad(a).rad(b).rad(c) < c
    if 5) ok test if gcd(a,b)=1
    if 5) and 6) ok then sum the current c.

The answer for this problem comes in 34 s.
There is still more room for improvement when I feel inspired.