Create Resolution of Problem 128 of Project Euler
For this problem I first thought of recreating this kind of structure in a matrix, and I had actually made it, also I used a Sieve of Eratosthenes for the prime issue, but the thing was that, for my surprise, after the first 200th tile the program would go slow and out of memory after the 600th tile. This is the program that was my first approach to this problem:

After that, when I get time and my head able to focus again on this problem, first I used the 
first 50 th tiles calculated from my first program, and tried to see if there was some kind of 
numerical pattern using various statistical methods, but I went nowhere. 
My second idea, already going out of clue of how to solve this, was to mark in the hexagonal
 matrix the solutions and see if there was some king of pattern. 
Well for my surprise, there was indeed, the center column, and the column right of center 
column were where the solutions were to be found. 
I had no clue why but started to work on this track and first tried to see on the paper if I 
could work on those columns without calculating the entire matrix. 
I was able to found the formulas so I advanced to a program: 

For the center column I used an array with ringHead[n-1], ringHead[n] and ringHead[n+1] 
for n being the ring number n. 

For the column right of the center column, I used an array with: lasElement[n-1], lastElement[n], lastElement[n+1] for n being the ring number n. 
For the dimension of the rings I noticed the following formula: 
dim(n) = 1 if n = 1 dim(n) = 6.(n-1) if n >= 2 

There are the 6 neighbors to analyze, and 1 special case in each column. 
In the center column, the special case is when we are in number 1, then neighbors are 2,3,4,5,6 
in the other case there are: 
ringHead[n+1] 
ringHead[n+1]+1 
ringHead[n]+1 
ringHead[n-1] 
lastElement[n] 
lastElement[n+1] 

In the column right from the center column, the special case is when we are in the number 7,
 on this case the neighbors are: 
ringHead[n-1] 
ringHead[n] 
lastElement[n]-1 
lastElement[n+1] 
lastElement[n+1]-1 
lastElement[n+1]-2 

if it is another number, then it is the general case, so the formula is: 
lastElement[n]-1 
lastElement[n-1] 
ringHead[n-1] 
ringHead[n] 
lastElement[n+1] 
lastElement[n+1]-1 

In the end of each ring, the new values of the ringHead[n+1] and of the lasElement[n+1] are using
 the following formula: 

/* next value n+1 */ 
ringHead[n+1] = ringHead[n]+dimCalc(n); 
lastElement[n+1] = ringHead[n+1]+dimCalc(n+1)-1; 

When we found a new tile, the nAnswer++, when nAnswer = 2000 we end our program. 
This program found the solution in 0.553 s.